+++ classes/MyAdmin.php
@@ -909,16 +909,28 @@ class MyAdmin extends MyCommon
             ]
         );
         $htmlbody = $this->outputAdminBody(); // MUST precede outputBodyEndInlineScript method so that $this->tableAdmin->script is already populated
+        $switches = [];
+        // test for presence of a string in_array is simpler than testing for existance and value of a boolean array
+        foreach (['change-password', 'create-user', 'delete-user', 'logout', 'media', 'user'] as $switch) {
+            // todo $_GET be replaced by an object property
+            if (isset($_GET[$switch])) {
+                $switches[] = $switch;
+            }
+        }
         $params = [
             'authUser' => (int) (isset($_SESSION['user']) && $_SESSION['user']), // 0 vs 1
             'clientSideResources' => $this->clientSideResources,
             'inlineJavaScript' => $this->outputBodyEndInlineScript(),
+            'featureFlags' => $this->featureFlags,
             'htmlbody' => $htmlbody,
             //'htmlhead' => $this->outputHead($this->getPageTitle()),
             'HTMLHeaders' => $this->HTMLHeaders,
             'language' => Tools::h($_SESSION['language']),
             'pageTitle' => $this->getPageTitle(),
+            'switches' => $switches,
             'token' => end($_SESSION['token']), // for login
+            'translations' => $this->tableAdmin->TRANSLATIONS, // languages for which translations are available
+            'username' => (isset($_SESSION['user']) && $_SESSION['user']) ? $_SESSION['user'] : null,
         ];
         $customFilters = new MyCustomFilters($this->MyCMS, [$this->tableAdmin, 'translate']);
         $render = new Render($this->template, DIR_TEMPLATE_CACHE, [$customFilters, 'common']);